@@ -1,14 +1,16 @@
-from typing import Optional, Union
+from typing import Any, Generic, TypeVar
 
+import gymnasium as gym
 import numpy as np
-from gym import Env, Space
-from gym.spaces import Box, Discrete, MultiBinary, MultiDiscrete
+from gymnasium import spaces
 
-from stable_baselines3.common.type_aliases import GymObs, GymStepReturn
+from stable_baselines3.common.type_aliases import GymStepReturn
+
+T = TypeVar("T", int, np.ndarray)
 
 
-class IdentityEnv(Env):
-    def __init__(self, dim: Optional[int] = None, space: Optional[Space] = None, ep_length: int = 100):
+class IdentityEnv(gym.Env, Generic[T]):
+    def __init__(self, dim: int | None = None, space: spaces.Space | None = None, ep_length: int = 100):
         """
         Identity environment for testing purposes
 
@@ -22,7 +24,7 @@ class IdentityEnv(Env):
         if space is None:
             if dim is None:
                 dim = 1
-            space = Discrete(dim)
+            space = spaces.Discrete(dim)
         else:
             assert dim is None, "arguments for both 'dim' and 'space' provided: at most one allowed"
 
@@ -32,30 +34,33 @@ class IdentityEnv(Env):
         self.num_resets = -1  # Becomes 0 after __init__ exits.
         self.reset()
 
-    def reset(self) -> GymObs:
+    def reset(self, *, seed: int | None = None, options: dict | None = None) -> tuple[T, dict]:
+        if seed is not None:
+            super().reset(seed=seed)
         self.current_step = 0
         self.num_resets += 1
         self._choose_next_state()
-        return self.state
+        return self.state, {}
 
-    def step(self, action: Union[int, np.ndarray]) -> GymStepReturn:
+    def step(self, action: T) -> tuple[T, float, bool, bool, dict[str, Any]]:
         reward = self._get_reward(action)
         self._choose_next_state()
         self.current_step += 1
-        done = self.current_step >= self.ep_length
-        return self.state, reward, done, {}
+        terminated = False
+        truncated = self.current_step >= self.ep_length
+        return self.state, reward, terminated, truncated, {}
 
     def _choose_next_state(self) -> None:
         self.state = self.action_space.sample()
 
-    def _get_reward(self, action: Union[int, np.ndarray]) -> float:
+    def _get_reward(self, action: T) -> float:
         return 1.0 if np.all(self.state == action) else 0.0
 
     def render(self, mode: str = "human") -> None:
         pass
 
 
-class IdentityEnvBox(IdentityEnv):
+class IdentityEnvBox(IdentityEnv[np.ndarray]):
     def __init__(self, low: float = -1.0, high: float = 1.0, eps: float = 0.05, ep_length: int = 100):
         """
         Identity environment for testing purposes
@@ -65,46 +70,47 @@ class IdentityEnvBox(IdentityEnv):
         :param eps: the epsilon bound for correct value
         :param ep_length: the length of each episode in timesteps
         """
-        space = Box(low=low, high=high, shape=(1,), dtype=np.float32)
+        space = spaces.Box(low=low, high=high, shape=(1,), dtype=np.float32)
         super().__init__(ep_length=ep_length, space=space)
         self.eps = eps
 
-    def step(self, action: np.ndarray) -> GymStepReturn:
+    def step(self, action: np.ndarray) -> tuple[np.ndarray, float, bool, bool, dict[str, Any]]:
         reward = self._get_reward(action)
         self._choose_next_state()
         self.current_step += 1
-        done = self.current_step >= self.ep_length
-        return self.state, reward, done, {}
+        terminated = False
+        truncated = self.current_step >= self.ep_length
+        return self.state, reward, terminated, truncated, {}
 
     def _get_reward(self, action: np.ndarray) -> float:
         return 1.0 if (self.state - self.eps) <= action <= (self.state + self.eps) else 0.0
 
 
-class IdentityEnvMultiDiscrete(IdentityEnv):
-    def __init__(self, dim: int = 1, ep_length: int = 100):
+class IdentityEnvMultiDiscrete(IdentityEnv[np.ndarray]):
+    def __init__(self, dim: int = 1, ep_length: int = 100) -> None:
         """
         Identity environment for testing purposes
 
         :param dim: the size of the dimensions you want to learn
         :param ep_length: the length of each episode in timesteps
         """
-        space = MultiDiscrete([dim, dim])
+        space = spaces.MultiDiscrete([dim, dim])
         super().__init__(ep_length=ep_length, space=space)
 
 
-class IdentityEnvMultiBinary(IdentityEnv):
-    def __init__(self, dim: int = 1, ep_length: int = 100):
+class IdentityEnvMultiBinary(IdentityEnv[np.ndarray]):
+    def __init__(self, dim: int = 1, ep_length: int = 100) -> None:
         """
         Identity environment for testing purposes
 
         :param dim: the size of the dimensions you want to learn
         :param ep_length: the length of each episode in timesteps
         """
-        space = MultiBinary(dim)
+        space = spaces.MultiBinary(dim)
         super().__init__(ep_length=ep_length, space=space)
 
 
-class FakeImageEnv(Env):
+class FakeImageEnv(gym.Env):
     """
     Fake image environment for testing purposes, it mimics Atari games.
 
@@ -124,27 +130,30 @@ class FakeImageEnv(Env):
         n_channels: int = 1,
         discrete: bool = True,
         channel_first: bool = False,
-    ):
+    ) -> None:
         self.observation_shape = (screen_height, screen_width, n_channels)
         if channel_first:
             self.observation_shape = (n_channels, screen_height, screen_width)
-        self.observation_space = Box(low=0, high=255, shape=self.observation_shape, dtype=np.uint8)
+        self.observation_space = spaces.Box(low=0, high=255, shape=self.observation_shape, dtype=np.uint8)
         if discrete:
-            self.action_space = Discrete(action_dim)
+            self.action_space = spaces.Discrete(action_dim)
         else:
-            self.action_space = Box(low=-1, high=1, shape=(5,), dtype=np.float32)
+            self.action_space = spaces.Box(low=-1, high=1, shape=(5,), dtype=np.float32)
         self.ep_length = 10
         self.current_step = 0
 
-    def reset(self) -> np.ndarray:
+    def reset(self, *, seed: int | None = None, options: dict | None = None) -> tuple[np.ndarray, dict]:
+        if seed is not None:
+            super().reset(seed=seed)
         self.current_step = 0
-        return self.observation_space.sample()
+        return self.observation_space.sample(), {}
 
-    def step(self, action: Union[np.ndarray, int]) -> GymStepReturn:
+    def step(self, action: np.ndarray | int) -> GymStepReturn:
         reward = 0.0
         self.current_step += 1
-        done = self.current_step >= self.ep_length
-        return self.observation_space.sample(), reward, done, {}
+        terminated = False
+        truncated = self.current_step >= self.ep_length
+        return self.observation_space.sample(), reward, terminated, truncated, {}
 
     def render(self, mode: str = "human") -> None:
         pass
