@@ -1,7 +1,6 @@
-from typing import Dict, Union
-
-import gym
+import gymnasium as gym
 import numpy as np
+from gymnasium import spaces
 
 from stable_baselines3.common.type_aliases import GymStepReturn
 
@@ -42,7 +41,7 @@ class SimpleMultiObsEnv(gym.Env):
         discrete_actions: bool = True,
         channel_last: bool = True,
     ):
-        super(SimpleMultiObsEnv, self).__init__()
+        super().__init__()
 
         self.vector_size = 5
         if channel_last:
@@ -53,14 +52,14 @@ class SimpleMultiObsEnv(gym.Env):
         self.random_start = random_start
         self.discrete_actions = discrete_actions
         if discrete_actions:
-            self.action_space = gym.spaces.Discrete(4)
+            self.action_space = spaces.Discrete(4)
         else:
-            self.action_space = gym.spaces.Box(0, 1, (4,))
+            self.action_space = spaces.Box(0, 1, (4,))
 
-        self.observation_space = gym.spaces.Dict(
+        self.observation_space = spaces.Dict(
             spaces={
-                "vec": gym.spaces.Box(0, 1, (self.vector_size,), dtype=np.float64),
-                "img": gym.spaces.Box(0, 255, self.img_size, dtype=np.uint8),
+                "vec": spaces.Box(0, 1, (self.vector_size,), dtype=np.float64),
+                "img": spaces.Box(0, 255, self.img_size, dtype=np.uint8),
             }
         )
         self.count = 0
@@ -72,7 +71,7 @@ class SimpleMultiObsEnv(gym.Env):
         self.init_possible_transitions()
 
         self.num_col = num_col
-        self.state_mapping = []
+        self.state_mapping: list[dict[str, np.ndarray]] = []
         self.init_state_mapping(num_col, num_row)
 
         self.max_state = len(self.state_mapping) - 1
@@ -93,7 +92,7 @@ class SimpleMultiObsEnv(gym.Env):
             for j in range(num_row):
                 self.state_mapping.append({"vec": col_vecs[i], "img": row_imgs[j].reshape(self.img_size)})
 
-    def get_state_mapping(self) -> Dict[str, np.ndarray]:
+    def get_state_mapping(self) -> dict[str, np.ndarray]:
         """
         Uses the state to get the observation mapping.
 
@@ -120,20 +119,18 @@ class SimpleMultiObsEnv(gym.Env):
         self.right_possible = [0, 1, 2, 12, 13, 14]
         self.up_possible = [4, 8, 12, 7, 11, 15]
 
-    def step(self, action: Union[int, float, np.ndarray]) -> GymStepReturn:
+    def step(self, action: int | np.ndarray) -> GymStepReturn:
         """
         Run one timestep of the environment's dynamics. When end of
         episode is reached, you are responsible for calling `reset()`
         to reset this environment's state.
-        Accepts an action and returns a tuple (observation, reward, done, info).
+        Accepts an action and returns a tuple (observation, reward, terminated, truncated, info).
 
         :param action:
-        :return: tuple (observation, reward, done, info).
+        :return: tuple (observation, reward, terminated, truncated, info).
         """
         if not self.discrete_actions:
-            action = np.argmax(action)
-        else:
-            action = int(action)
+            action = np.argmax(action)  # type: ignore[assignment]
 
         self.count += 1
 
@@ -151,12 +148,13 @@ class SimpleMultiObsEnv(gym.Env):
             self.state -= self.num_col
 
         got_to_end = self.state == self.max_state
-        reward = 1 if got_to_end else reward
-        done = self.count > self.max_count or got_to_end
+        reward = 1.0 if got_to_end else reward
+        truncated = self.count > self.max_count
+        terminated = got_to_end
 
         self.log = f"Went {self.action2str[action]} in state {prev_state}, got to state {self.state}"
 
-        return self.get_state_mapping(), reward, done, {"got_to_end": got_to_end}
+        return self.get_state_mapping(), reward, terminated, truncated, {"got_to_end": got_to_end}
 
     def render(self, mode: str = "human") -> None:
         """
@@ -166,15 +164,18 @@ class SimpleMultiObsEnv(gym.Env):
         """
         print(self.log)
 
-    def reset(self) -> Dict[str, np.ndarray]:
+    def reset(self, *, seed: int | None = None, options: dict | None = None) -> tuple[dict[str, np.ndarray], dict]:
         """
         Resets the environment state and step count and returns reset observation.
 
+        :param seed:
         :return: observation dict {'vec': ..., 'img': ...}
         """
+        if seed is not None:
+            super().reset(seed=seed)
         self.count = 0
         if not self.random_start:
             self.state = 0
         else:
             self.state = np.random.randint(0, self.max_state)
-        return self.state_mapping[self.state]
+        return self.state_mapping[self.state], {}
