@@ -5,16 +5,14 @@ import json
 import os
 import time
 from glob import glob
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Any, SupportsFloat
 
-import gym
-import numpy as np
+import gymnasium as gym
 import pandas
+from gymnasium.core import ActType, ObsType
 
-from stable_baselines3.common.type_aliases import GymObs, GymStepReturn
 
-
-class Monitor(gym.Wrapper):
+class Monitor(gym.Wrapper[ObsType, ActType, ObsType, ActType]):
     """
     A monitor wrapper for Gym environments, it is used to know the episode reward, length, time and other data.
 
@@ -24,6 +22,8 @@ class Monitor(gym.Wrapper):
     :param reset_keywords: extra keywords for the reset call,
         if extra parameters are needed at reset
     :param info_keywords: extra information to log, from the information return of env.step()
+    :param override_existing: appends to file if ``filename`` exists, otherwise
+        override existing files (default)
     """
 
     EXT = "monitor.csv"
@@ -31,33 +31,37 @@ class Monitor(gym.Wrapper):
     def __init__(
         self,
         env: gym.Env,
-        filename: Optional[str] = None,
+        filename: str | None = None,
         allow_early_resets: bool = True,
-        reset_keywords: Tuple[str, ...] = (),
-        info_keywords: Tuple[str, ...] = (),
+        reset_keywords: tuple[str, ...] = (),
+        info_keywords: tuple[str, ...] = (),
+        override_existing: bool = True,
     ):
-        super(Monitor, self).__init__(env=env)
+        super().__init__(env=env)
         self.t_start = time.time()
+        self.results_writer = None
         if filename is not None:
+            env_id = env.spec.id if env.spec is not None else None
             self.results_writer = ResultsWriter(
                 filename,
-                header={"t_start": self.t_start, "env_id": env.spec and env.spec.id},
+                header={"t_start": self.t_start, "env_id": str(env_id)},
                 extra_keys=reset_keywords + info_keywords,
+                override_existing=override_existing,
             )
-        else:
-            self.results_writer = None
+
         self.reset_keywords = reset_keywords
         self.info_keywords = info_keywords
         self.allow_early_resets = allow_early_resets
-        self.rewards = None
+        self.rewards: list[float] = []
         self.needs_reset = True
-        self.episode_returns = []
-        self.episode_lengths = []
-        self.episode_times = []
+        self.episode_returns: list[float] = []
+        self.episode_lengths: list[int] = []
+        self.episode_times: list[float] = []
         self.total_steps = 0
-        self.current_reset_info = {}  # extra info about the current episode, that was passed in during reset()
+        # extra info about the current episode, that was passed in during reset()
+        self.current_reset_info: dict[str, Any] = {}
 
-    def reset(self, **kwargs) -> GymObs:
+    def reset(self, **kwargs) -> tuple[ObsType, dict[str, Any]]:
         """
         Calls the Gym environment reset. Can only be called if the environment is over, or if allow_early_resets is True
 
@@ -78,18 +82,18 @@ class Monitor(gym.Wrapper):
             self.current_reset_info[key] = value
         return self.env.reset(**kwargs)
 
-    def step(self, action: Union[np.ndarray, int]) -> GymStepReturn:
+    def step(self, action: ActType) -> tuple[ObsType, SupportsFloat, bool, bool, dict[str, Any]]:
         """
         Step the environment with the given action
 
         :param action: the action
-        :return: observation, reward, done, information
+        :return: observation, reward, terminated, truncated, information
         """
         if self.needs_reset:
             raise RuntimeError("Tried to step environment that needs reset")
-        observation, reward, done, info = self.env.step(action)
-        self.rewards.append(reward)
-        if done:
+        observation, reward, terminated, truncated, info = self.env.step(action)
+        self.rewards.append(float(reward))
+        if terminated or truncated:
             self.needs_reset = True
             ep_rew = sum(self.rewards)
             ep_len = len(self.rewards)
@@ -104,13 +108,13 @@ class Monitor(gym.Wrapper):
                 self.results_writer.write_row(ep_info)
             info["episode"] = ep_info
         self.total_steps += 1
-        return observation, reward, done, info
+        return observation, reward, terminated, truncated, info
 
     def close(self) -> None:
         """
         Closes the environment
         """
-        super(Monitor, self).close()
+        super().close()
         if self.results_writer is not None:
             self.results_writer.close()
 
@@ -122,7 +126,7 @@ class Monitor(gym.Wrapper):
         """
         return self.total_steps
 
-    def get_episode_rewards(self) -> List[float]:
+    def get_episode_rewards(self) -> list[float]:
         """
         Returns the rewards of all the episodes
 
@@ -130,7 +134,7 @@ class Monitor(gym.Wrapper):
         """
         return self.episode_returns
 
-    def get_episode_lengths(self) -> List[int]:
+    def get_episode_lengths(self) -> list[int]:
         """
         Returns the number of timesteps of all the episodes
 
@@ -138,7 +142,7 @@ class Monitor(gym.Wrapper):
         """
         return self.episode_lengths
 
-    def get_episode_times(self) -> List[float]:
+    def get_episode_times(self) -> list[float]:
         """
         Returns the runtime in seconds of all the episodes
 
@@ -159,17 +163,21 @@ class ResultsWriter:
     """
     A result writer that saves the data from the `Monitor` class
 
-    :param filename: the location to save a log file, can be None for no log
+    :param filename: the location to save a log file. When it does not end in
+        the string ``"monitor.csv"``, this suffix will be appended to it
     :param header: the header dictionary object of the saved csv
-    :param reset_keywords: the extra information to log, typically is composed of
+    :param extra_keys: the extra information to log, typically is composed of
         ``reset_keywords`` and ``info_keywords``
+    :param override_existing: appends to file if ``filename`` exists, otherwise
+        override existing files (default)
     """
 
     def __init__(
         self,
         filename: str = "",
-        header: Optional[Dict[str, Union[float, str]]] = None,
-        extra_keys: Tuple[str, ...] = (),
+        header: dict[str, float | str] | None = None,
+        extra_keys: tuple[str, ...] = (),
+        override_existing: bool = True,
     ):
         if header is None:
             header = {}
@@ -178,16 +186,23 @@ class ResultsWriter:
                 filename = os.path.join(filename, Monitor.EXT)
             else:
                 filename = filename + "." + Monitor.EXT
+        filename = os.path.realpath(filename)
+        # Create (if any) missing filename directories
+        os.makedirs(os.path.dirname(filename), exist_ok=True)
+        # Append mode when not overriding existing file
+        mode = "w" if override_existing else "a"
         # Prevent newline issue on Windows, see GH issue #692
-        self.file_handler = open(filename, "wt", newline="\n")
-        self.file_handler.write("#%s\n" % json.dumps(header))
-        self.logger = csv.DictWriter(self.file_handler, fieldnames=("r", "l", "t") + extra_keys)
-        self.logger.writeheader()
+        self.file_handler = open(filename, f"{mode}t", newline="\n")
+        self.logger = csv.DictWriter(self.file_handler, fieldnames=("r", "l", "t", *extra_keys))
+        if override_existing:
+            self.file_handler.write(f"#{json.dumps(header)}\n")
+            self.logger.writeheader()
+
         self.file_handler.flush()
 
-    def write_row(self, epinfo: Dict[str, Union[float, int]]) -> None:
+    def write_row(self, epinfo: dict[str, float]) -> None:
         """
-        Close the file handler
+        Write row of monitor data to csv log file.
 
         :param epinfo: the information on episodic return, length, and time
         """
@@ -202,7 +217,7 @@ class ResultsWriter:
         self.file_handler.close()
 
 
-def get_monitor_files(path: str) -> List[str]:
+def get_monitor_files(path: str) -> list[str]:
     """
     get all the monitor files in the given path
 
@@ -224,7 +239,7 @@ def load_results(path: str) -> pandas.DataFrame:
         raise LoadMonitorResultsError(f"No monitor files of the form *{Monitor.EXT} found in {path}")
     data_frames, headers = [], []
     for file_name in monitor_files:
-        with open(file_name, "rt") as file_handler:
+        with open(file_name) as file_handler:
             first_line = file_handler.readline()
             assert first_line[0] == "#"
             header = json.loads(first_line[1:])
@@ -232,6 +247,13 @@ def load_results(path: str) -> pandas.DataFrame:
             headers.append(header)
             data_frame["t"] += header["t_start"]
         data_frames.append(data_frame)
+    data_frames = [df for df in data_frames if not df.empty]
+    if not data_frames:
+        # Only empty monitor files, return empty df
+        empty_df = pandas.DataFrame(columns=["r", "l", "t"])
+        # Create index to have the same columns
+        empty_df.reset_index(inplace=True)
+        return empty_df
     data_frame = pandas.concat(data_frames)
     data_frame.sort_values("t", inplace=True)
     data_frame.reset_index(inplace=True)
