@@ -1,23 +1,31 @@
 """Common aliases for type hints"""
 
+from collections.abc import Callable
 from enum import Enum
-from typing import Any, Callable, Dict, List, NamedTuple, Tuple, Union
+from typing import TYPE_CHECKING, Any, NamedTuple, Protocol, SupportsFloat, Union
 
-import gym
+import gymnasium as gym
 import numpy as np
 import torch as th
 
-from stable_baselines3.common import callbacks, vec_env
+# Avoid circular imports, we use type hint as string to avoid it too
+if TYPE_CHECKING:
+    from stable_baselines3.common.callbacks import BaseCallback
+    from stable_baselines3.common.vec_env import VecEnv
 
-GymEnv = Union[gym.Env, vec_env.VecEnv]
-GymObs = Union[Tuple, Dict[str, Any], np.ndarray, int]
-GymStepReturn = Tuple[GymObs, float, bool, Dict]
-TensorDict = Dict[Union[str, int], th.Tensor]
-OptimizerStateDict = Dict[str, Any]
-MaybeCallback = Union[None, Callable, List[callbacks.BaseCallback], callbacks.BaseCallback]
+GymEnv = Union[gym.Env, "VecEnv"]
+GymObs = Union[tuple, dict[str, Any], np.ndarray, int]  # noqa: UP007
+GymResetReturn = tuple[GymObs, dict]
+AtariResetReturn = tuple[np.ndarray, dict[str, Any]]
+GymStepReturn = tuple[GymObs, float, bool, bool, dict]
+AtariStepReturn = tuple[np.ndarray, SupportsFloat, bool, bool, dict[str, Any]]
+TensorDict = dict[str, th.Tensor]
+OptimizerStateDict = dict[str, Any]
+MaybeCallback = Union[None, Callable, list["BaseCallback"], "BaseCallback"]
+PyTorchObs = Union[th.Tensor, TensorDict]  # noqa: UP007
 
 # A schedule takes the remaining progress as input
-# and ouputs a scalar (e.g. learning rate, clip range, ...)
+# and outputs a scalar (e.g. learning rate, clip range, ...)
 Schedule = Callable[[float], float]
 
 
@@ -30,7 +38,7 @@ class RolloutBufferSamples(NamedTuple):
     returns: th.Tensor
 
 
-class DictRolloutBufferSamples(RolloutBufferSamples):
+class DictRolloutBufferSamples(NamedTuple):
     observations: TensorDict
     actions: th.Tensor
     old_values: th.Tensor
@@ -45,14 +53,17 @@ class ReplayBufferSamples(NamedTuple):
     next_observations: th.Tensor
     dones: th.Tensor
     rewards: th.Tensor
+    # For n-step replay buffer
+    discounts: th.Tensor | None = None
 
 
-class DictReplayBufferSamples(ReplayBufferSamples):
+class DictReplayBufferSamples(NamedTuple):
     observations: TensorDict
     actions: th.Tensor
-    next_observations: th.Tensor
+    next_observations: TensorDict
     dones: th.Tensor
     rewards: th.Tensor
+    discounts: th.Tensor | None = None
 
 
 class RolloutReturn(NamedTuple):
@@ -69,3 +80,26 @@ class TrainFrequencyUnit(Enum):
 class TrainFreq(NamedTuple):
     frequency: int
     unit: TrainFrequencyUnit  # either "step" or "episode"
+
+
+class PolicyPredictor(Protocol):
+    def predict(
+        self,
+        observation: np.ndarray | dict[str, np.ndarray],
+        state: tuple[np.ndarray, ...] | None = None,
+        episode_start: np.ndarray | None = None,
+        deterministic: bool = False,
+    ) -> tuple[np.ndarray, tuple[np.ndarray, ...] | None]:
+        """
+        Get the policy action from an observation (and optional hidden state).
+        Includes sugar-coating to handle different observations (e.g. normalizing images).
+
+        :param observation: the input observation
+        :param state: The last hidden states (can be None, used in recurrent policies)
+        :param episode_start: The last masks (can be None, used in recurrent policies)
+            this correspond to beginning of episodes,
+            where the hidden states of the RNN must be reset.
+        :param deterministic: Whether or not to return deterministic actions.
+        :return: the model's action and the next hidden state
+            (used in recurrent policies)
+        """
