@@ -1,12 +1,13 @@
+import inspect
 import pickle
-import warnings
 from copy import deepcopy
-from typing import Any, Dict, List, Optional, Union
+from typing import Any
 
-import gym
 import numpy as np
+from gymnasium import spaces
 
 from stable_baselines3.common import utils
+from stable_baselines3.common.preprocessing import is_image_space
 from stable_baselines3.common.running_mean_std import RunningMeanStd
 from stable_baselines3.common.vec_env.base_vec_env import VecEnv, VecEnvStepReturn, VecEnvWrapper
 
@@ -28,6 +29,9 @@ class VecNormalize(VecEnvWrapper):
         If not specified, all keys will be normalized.
     """
 
+    obs_spaces: dict[str, spaces.Space]
+    old_obs: np.ndarray | dict[str, np.ndarray]
+
     def __init__(
         self,
         venv: VecEnv,
@@ -38,7 +42,7 @@ class VecNormalize(VecEnvWrapper):
         clip_reward: float = 10.0,
         gamma: float = 0.99,
         epsilon: float = 1e-8,
-        norm_obs_keys: Optional[List[str]] = None,
+        norm_obs_keys: list[str] | None = None,
     ):
         VecEnvWrapper.__init__(self, venv)
 
@@ -46,14 +50,40 @@ class VecNormalize(VecEnvWrapper):
         self.norm_obs_keys = norm_obs_keys
         # Check observation spaces
         if self.norm_obs:
+            # Note: mypy doesn't take into account the sanity checks, which lead to several type: ignore...
             self._sanity_checks()
 
-        if isinstance(self.observation_space, gym.spaces.Dict):
-            self.obs_spaces = self.observation_space.spaces
-            self.obs_rms = {key: RunningMeanStd(shape=self.obs_spaces[key].shape) for key in self.norm_obs_keys}
-        else:
-            self.obs_spaces = None
-            self.obs_rms = RunningMeanStd(shape=self.observation_space.shape)
+            if isinstance(self.observation_space, spaces.Dict):
+                self.obs_spaces = self.observation_space.spaces
+                self.obs_rms = {key: RunningMeanStd(shape=self.obs_spaces[key].shape) for key in self.norm_obs_keys}  # type: ignore[arg-type, union-attr]
+                # Update observation space when using image
+                # See explanation below and GH #1214
+                for key in self.obs_rms.keys():
+                    if is_image_space(self.obs_spaces[key]):
+                        self.observation_space.spaces[key] = spaces.Box(
+                            low=-clip_obs,
+                            high=clip_obs,
+                            shape=self.obs_spaces[key].shape,
+                            dtype=np.float32,
+                        )
+
+            else:
+                self.obs_rms = RunningMeanStd(shape=self.observation_space.shape)  # type: ignore[assignment, arg-type]
+                # Update observation space when using image
+                # See GH #1214
+                # This is to raise proper error when
+                # VecNormalize is used with an image-like input and
+                # normalize_images=True.
+                # For correctness, we should also update the bounds
+                # in other cases but this will cause backward-incompatible change
+                # and break already saved policies.
+                if is_image_space(self.observation_space):
+                    self.observation_space = spaces.Box(
+                        low=-clip_obs,
+                        high=clip_obs,
+                        shape=self.observation_space.shape,
+                        dtype=np.float32,
+                    )
 
         self.ret_rms = RunningMeanStd(shape=())
         self.clip_obs = clip_obs
@@ -65,28 +95,27 @@ class VecNormalize(VecEnvWrapper):
         self.training = training
         self.norm_obs = norm_obs
         self.norm_reward = norm_reward
-        self.old_obs = np.array([])
         self.old_reward = np.array([])
 
     def _sanity_checks(self) -> None:
         """
         Check the observations that are going to be normalized are of the correct type (spaces.Box).
         """
-        if isinstance(self.observation_space, gym.spaces.Dict):
+        if isinstance(self.observation_space, spaces.Dict):
             # By default, we normalize all keys
             if self.norm_obs_keys is None:
                 self.norm_obs_keys = list(self.observation_space.spaces.keys())
             # Check that all keys are of type Box
             for obs_key in self.norm_obs_keys:
-                if not isinstance(self.observation_space.spaces[obs_key], gym.spaces.Box):
+                if not isinstance(self.observation_space.spaces[obs_key], spaces.Box):
                     raise ValueError(
                         f"VecNormalize only supports `gym.spaces.Box` observation spaces but {obs_key} "
                         f"is of type {self.observation_space.spaces[obs_key]}. "
-                        "You should probably explicitely pass the observation keys "
+                        "You should probably explicitly pass the observation keys "
                         " that should be normalized via the `norm_obs_keys` parameter."
                     )
 
-        elif isinstance(self.observation_space, gym.spaces.Box):
+        elif isinstance(self.observation_space, spaces.Box):
             if self.norm_obs_keys is not None:
                 raise ValueError("`norm_obs_keys` param is applicable only with `gym.spaces.Dict` observation spaces")
 
@@ -96,7 +125,7 @@ class VecNormalize(VecEnvWrapper):
                 f"not {self.observation_space}"
             )
 
-    def __getstate__(self) -> Dict[str, Any]:
+    def __getstate__(self) -> dict[str, Any]:
         """
         Gets state for pickling.
 
@@ -109,7 +138,7 @@ class VecNormalize(VecEnvWrapper):
         del state["returns"]
         return state
 
-    def __setstate__(self, state: Dict[str, Any]) -> None:
+    def __setstate__(self, state: dict[str, Any]) -> None:
         """
         Restores pickled state.
 
@@ -117,11 +146,11 @@ class VecNormalize(VecEnvWrapper):
 
         :param state:"""
         # Backward compatibility
-        if "norm_obs_keys" not in state and isinstance(state["observation_space"], gym.spaces.Dict):
+        if "norm_obs_keys" not in state and isinstance(state["observation_space"], spaces.Dict):
             state["norm_obs_keys"] = list(state["observation_space"].spaces.keys())
         self.__dict__.update(state)
         assert "venv" not in state
-        self.venv = None
+        self.venv = None  # type: ignore[assignment]
 
     def set_venv(self, venv: VecEnv) -> None:
         """
@@ -133,10 +162,13 @@ class VecNormalize(VecEnvWrapper):
         """
         if self.venv is not None:
             raise ValueError("Trying to set venv of already initialized VecNormalize wrapper.")
-        VecEnvWrapper.__init__(self, venv)
+        self.venv = venv
+        self.num_envs = venv.num_envs
+        self.class_attributes = dict(inspect.getmembers(self.__class__))
+        self.render_mode = venv.render_mode
 
-        # Check only that the observation_space match
-        utils.check_for_correct_spaces(venv, self.observation_space, venv.action_space)
+        # Check that the observation_space shape match
+        utils.check_shape_equal(self.observation_space, venv.observation_space)
         self.returns = np.zeros(self.num_envs)
 
     def step_wait(self) -> VecEnvStepReturn:
@@ -147,10 +179,11 @@ class VecNormalize(VecEnvWrapper):
         where ``dones`` is a boolean vector indicating whether each element is new.
         """
         obs, rewards, dones, infos = self.venv.step_wait()
+        assert isinstance(obs, (np.ndarray, dict))  # for mypy
         self.old_obs = obs
         self.old_reward = rewards
 
-        if self.training:
+        if self.training and self.norm_obs:
             if isinstance(obs, dict) and isinstance(self.obs_rms, dict):
                 for key in self.obs_rms.keys():
                     self.obs_rms[key].update(obs[key])
@@ -196,7 +229,7 @@ class VecNormalize(VecEnvWrapper):
         """
         return (obs * np.sqrt(obs_rms.var + self.epsilon)) + obs_rms.mean
 
-    def normalize_obs(self, obs: Union[np.ndarray, Dict[str, np.ndarray]]) -> Union[np.ndarray, Dict[str, np.ndarray]]:
+    def normalize_obs(self, obs: np.ndarray | dict[str, np.ndarray]) -> np.ndarray | dict[str, np.ndarray]:
         """
         Normalize observations using this VecNormalize's observations statistics.
         Calling this method does not update statistics.
@@ -205,10 +238,12 @@ class VecNormalize(VecEnvWrapper):
         obs_ = deepcopy(obs)
         if self.norm_obs:
             if isinstance(obs, dict) and isinstance(self.obs_rms, dict):
+                assert self.norm_obs_keys is not None
                 # Only normalize the specified keys
                 for key in self.norm_obs_keys:
-                    obs_[key] = self._normalize_obs(obs[key], self.obs_rms[key]).astype(np.float32)
+                    obs_[key] = self._normalize_obs(obs[key], self.obs_rms[key]).astype(np.float32)  # type: ignore[call-overload]
             else:
+                assert isinstance(self.obs_rms, RunningMeanStd)
                 obs_ = self._normalize_obs(obs, self.obs_rms).astype(np.float32)
         return obs_
 
@@ -219,16 +254,20 @@ class VecNormalize(VecEnvWrapper):
         """
         if self.norm_reward:
             reward = np.clip(reward / np.sqrt(self.ret_rms.var + self.epsilon), -self.clip_reward, self.clip_reward)
-        return reward
+        # Note: we cast to float32 as it correspond to Python default float type
+        # This cast is needed because `RunningMeanStd` keeps stats in float64
+        return reward.astype(np.float32)
 
-    def unnormalize_obs(self, obs: Union[np.ndarray, Dict[str, np.ndarray]]) -> Union[np.ndarray, Dict[str, np.ndarray]]:
+    def unnormalize_obs(self, obs: np.ndarray | dict[str, np.ndarray]) -> np.ndarray | dict[str, np.ndarray]:
         # Avoid modifying by reference the original object
         obs_ = deepcopy(obs)
         if self.norm_obs:
             if isinstance(obs, dict) and isinstance(self.obs_rms, dict):
+                assert self.norm_obs_keys is not None
                 for key in self.norm_obs_keys:
-                    obs_[key] = self._unnormalize_obs(obs[key], self.obs_rms[key])
+                    obs_[key] = self._unnormalize_obs(obs[key], self.obs_rms[key])  # type: ignore[call-overload]
             else:
+                assert isinstance(self.obs_rms, RunningMeanStd)
                 obs_ = self._unnormalize_obs(obs, self.obs_rms)
         return obs_
 
@@ -237,7 +276,7 @@ class VecNormalize(VecEnvWrapper):
             return reward * np.sqrt(self.ret_rms.var + self.epsilon)
         return reward
 
-    def get_original_obs(self) -> Union[np.ndarray, Dict[str, np.ndarray]]:
+    def get_original_obs(self) -> np.ndarray | dict[str, np.ndarray]:
         """
         Returns an unnormalized version of the observations from the most recent
         step or reset.
@@ -250,19 +289,21 @@ class VecNormalize(VecEnvWrapper):
         """
         return self.old_reward.copy()
 
-    def reset(self) -> Union[np.ndarray, Dict[str, np.ndarray]]:
+    def reset(self) -> np.ndarray | dict[str, np.ndarray]:
         """
         Reset all environments
         :return: first observation of the episode
         """
         obs = self.venv.reset()
+        assert isinstance(obs, (np.ndarray, dict))
         self.old_obs = obs
         self.returns = np.zeros(self.num_envs)
-        if self.training:
+        if self.training and self.norm_obs:
             if isinstance(obs, dict) and isinstance(self.obs_rms, dict):
                 for key in self.obs_rms.keys():
                     self.obs_rms[key].update(obs[key])
             else:
+                assert isinstance(self.obs_rms, RunningMeanStd)
                 self.obs_rms.update(obs)
         return self.normalize_obs(obs)
 
@@ -289,8 +330,3 @@ class VecNormalize(VecEnvWrapper):
         """
         with open(save_path, "wb") as file_handler:
             pickle.dump(self, file_handler)
-
-    @property
-    def ret(self) -> np.ndarray:
-        warnings.warn("`VecNormalize` `ret` attribute is deprecated. Please use `returns` instead.", DeprecationWarning)
-        return self.returns
