@@ -1,8 +1,7 @@
 from copy import deepcopy
-from typing import Dict, Union
 
 import numpy as np
-from gym import spaces
+from gymnasium import spaces
 
 from stable_baselines3.common.preprocessing import is_image_space, is_image_space_channels_first
 from stable_baselines3.common.vec_env.base_vec_env import VecEnv, VecEnvStepReturn, VecEnvWrapper
@@ -20,26 +19,28 @@ class VecTransposeImage(VecEnvWrapper):
 
     def __init__(self, venv: VecEnv, skip: bool = False):
         assert is_image_space(venv.observation_space) or isinstance(
-            venv.observation_space, spaces.dict.Dict
+            venv.observation_space, spaces.Dict
         ), "The observation space must be an image or dictionary observation space"
 
         self.skip = skip
         # Do nothing
         if skip:
-            super(VecTransposeImage, self).__init__(venv)
+            super().__init__(venv)
             return
 
-        if isinstance(venv.observation_space, spaces.dict.Dict):
+        if isinstance(venv.observation_space, spaces.Dict):
             self.image_space_keys = []
             observation_space = deepcopy(venv.observation_space)
             for key, space in observation_space.spaces.items():
                 if is_image_space(space):
                     # Keep track of which keys should be transposed later
                     self.image_space_keys.append(key)
+                    assert isinstance(space, spaces.Box)
                     observation_space.spaces[key] = self.transpose_space(space, key)
         else:
-            observation_space = self.transpose_space(venv.observation_space)
-        super(VecTransposeImage, self).__init__(venv, observation_space=observation_space)
+            assert isinstance(venv.observation_space, spaces.Box)
+            observation_space = self.transpose_space(venv.observation_space)  # type: ignore[assignment]
+        super().__init__(venv, observation_space=observation_space)
 
     @staticmethod
     def transpose_space(observation_space: spaces.Box, key: str = "") -> spaces.Box:
@@ -57,7 +58,7 @@ class VecTransposeImage(VecEnvWrapper):
         ), f"The observation space {key} must follow the channel last convention"
         height, width, channels = observation_space.shape
         new_shape = (channels, height, width)
-        return spaces.Box(low=0, high=255, shape=new_shape, dtype=observation_space.dtype)
+        return spaces.Box(low=0, high=255, shape=new_shape, dtype=observation_space.dtype)  # type: ignore[arg-type]
 
     @staticmethod
     def transpose_image(image: np.ndarray) -> np.ndarray:
@@ -71,7 +72,7 @@ class VecTransposeImage(VecEnvWrapper):
             return np.transpose(image, (2, 0, 1))
         return np.transpose(image, (0, 3, 1, 2))
 
-    def transpose_observations(self, observations: Union[np.ndarray, Dict]) -> Union[np.ndarray, Dict]:
+    def transpose_observations(self, observations: np.ndarray | dict) -> np.ndarray | dict:
         """
         Transpose (if needed) and return new observations.
 
@@ -101,13 +102,16 @@ class VecTransposeImage(VecEnvWrapper):
             if "terminal_observation" in infos[idx]:
                 infos[idx]["terminal_observation"] = self.transpose_observations(infos[idx]["terminal_observation"])
 
+        assert isinstance(observations, (np.ndarray, dict))
         return self.transpose_observations(observations), rewards, dones, infos
 
-    def reset(self) -> Union[np.ndarray, Dict]:
+    def reset(self) -> np.ndarray | dict:
         """
         Reset all environments
         """
-        return self.transpose_observations(self.venv.reset())
+        observations = self.venv.reset()
+        assert isinstance(observations, (np.ndarray, dict))
+        return self.transpose_observations(observations)
 
     def close(self) -> None:
         self.venv.close()
