@@ -1,7 +1,9 @@
 import os
-from typing import Callable
+import os.path
+from collections.abc import Callable
 
-from gym.wrappers.monitoring import video_recorder
+import numpy as np
+from gymnasium import error, logger
 
 from stable_baselines3.common.vec_env.base_vec_env import VecEnv, VecEnvObs, VecEnvStepReturn, VecEnvWrapper
 from stable_baselines3.common.vec_env.dummy_vec_env import DummyVecEnv
@@ -13,6 +15,11 @@ class VecVideoRecorder(VecEnvWrapper):
     Wraps a VecEnv or VecEnvWrapper object to record rendered image as mp4 video.
     It requires ffmpeg or avconv to be installed on the machine.
 
+    Note: for now it only allows to record one video and all videos
+    must have at least two frames.
+
+    The video recorder code was adapted from Gymnasium v1.0.
+
     :param venv:
     :param video_folder: Where to save videos
     :param record_video_trigger: Function that defines when to start recording.
@@ -22,6 +29,9 @@ class VecVideoRecorder(VecEnvWrapper):
     :param name_prefix: Prefix to the video name
     """
 
+    video_name: str
+    video_path: str
+
     def __init__(
         self,
         venv: VecEnv,
@@ -30,7 +40,6 @@ class VecVideoRecorder(VecEnvWrapper):
         video_length: int = 200,
         name_prefix: str = "rl-video",
     ):
-
         VecEnvWrapper.__init__(self, venv)
 
         self.env = venv
@@ -44,14 +53,15 @@ class VecVideoRecorder(VecEnvWrapper):
 
         if isinstance(temp_env, DummyVecEnv) or isinstance(temp_env, SubprocVecEnv):
             metadata = temp_env.get_attr("metadata")[0]
-        else:
+        else:  # pragma: no cover # assume gym interface
             metadata = temp_env.metadata
 
         self.env.metadata = metadata
+        assert self.env.render_mode == "rgb_array", f"The render_mode must be 'rgb_array', not {self.env.render_mode}"
+
+        self.frames_per_sec = self.env.metadata.get("render_fps", 30)
 
         self.record_video_trigger = record_video_trigger
-        self.video_recorder = None
-
         self.video_folder = os.path.abspath(video_folder)
         # Create output folder if needed
         os.makedirs(self.video_folder, exist_ok=True)
@@ -61,53 +71,87 @@ class VecVideoRecorder(VecEnvWrapper):
         self.video_length = video_length
 
         self.recording = False
-        self.recorded_frames = 0
+        self.recorded_frames: list[np.ndarray] = []
+
+        try:
+            import moviepy  # noqa: F401
+        except ImportError as e:  # pragma: no cover
+            raise error.DependencyNotInstalled("MoviePy is not installed, run `pip install 'gymnasium[other]'`") from e
 
     def reset(self) -> VecEnvObs:
         obs = self.venv.reset()
-        self.start_video_recorder()
+        if self._video_enabled():
+            self._start_video_recorder()
         return obs
 
-    def start_video_recorder(self) -> None:
-        self.close_video_recorder()
-
-        video_name = f"{self.name_prefix}-step-{self.step_id}-to-step-{self.step_id + self.video_length}"
-        base_path = os.path.join(self.video_folder, video_name)
-        self.video_recorder = video_recorder.VideoRecorder(
-            env=self.env, base_path=base_path, metadata={"step_id": self.step_id}
-        )
-
-        self.video_recorder.capture_frame()
-        self.recorded_frames = 1
-        self.recording = True
+    def _start_video_recorder(self) -> None:
+        # Update video name and path
+        self.video_name = f"{self.name_prefix}-step-{self.step_id}-to-step-{self.step_id + self.video_length}.mp4"
+        self.video_path = os.path.join(self.video_folder, self.video_name)
+        self._start_recording()
+        self._capture_frame()
 
     def _video_enabled(self) -> bool:
         return self.record_video_trigger(self.step_id)
 
     def step_wait(self) -> VecEnvStepReturn:
-        obs, rews, dones, infos = self.venv.step_wait()
+        obs, rewards, dones, infos = self.venv.step_wait()
 
         self.step_id += 1
         if self.recording:
-            self.video_recorder.capture_frame()
-            self.recorded_frames += 1
-            if self.recorded_frames > self.video_length:
-                print(f"Saving video to {self.video_recorder.path}")
-                self.close_video_recorder()
+            self._capture_frame()
+            if len(self.recorded_frames) > self.video_length:
+                print(f"Saving video to {self.video_path}")
+                self._stop_recording()
         elif self._video_enabled():
-            self.start_video_recorder()
+            self._start_video_recorder()
 
-        return obs, rews, dones, infos
+        return obs, rewards, dones, infos
 
-    def close_video_recorder(self) -> None:
-        if self.recording:
-            self.video_recorder.close()
-        self.recording = False
-        self.recorded_frames = 1
+    def _capture_frame(self) -> None:
+        assert self.recording, "Cannot capture a frame, recording wasn't started."
+
+        frame = self.env.render()
+
+        if isinstance(frame, np.ndarray):
+            self.recorded_frames.append(frame)
+        else:
+            self._stop_recording()
+            logger.warn(
+                f"Recording stopped: expected type of frame returned by render to be a numpy array, got instead {type(frame)}."
+            )
 
     def close(self) -> None:
+        """Closes the wrapper then the video recorder."""
         VecEnvWrapper.close(self)
-        self.close_video_recorder()
+        if self.recording:  # pragma: no cover
+            self._stop_recording()
 
-    def __del__(self):
-        self.close()
+    def _start_recording(self) -> None:
+        """Start a new recording. If it is already recording, stops the current recording before starting the new one."""
+        if self.recording:  # pragma: no cover
+            self._stop_recording()
+
+        self.recording = True
+
+    def _stop_recording(self) -> None:
+        """Stop current recording and saves the video."""
+        assert self.recording, "_stop_recording was called, but no recording was started"
+
+        if len(self.recorded_frames) == 0:  # pragma: no cover
+            logger.warn("Ignored saving a video as there were zero frames to save.")
+        else:
+            from moviepy.video.io.ImageSequenceClip import ImageSequenceClip
+
+            clip = ImageSequenceClip(self.recorded_frames, fps=self.frames_per_sec)
+            clip.write_videofile(self.video_path)
+            del clip
+
+        del self.recorded_frames
+        self.recorded_frames = []
+        self.recording = False
+
+    def __del__(self) -> None:
+        """Warn the user in case last video wasn't saved."""
+        if len(self.recorded_frames) > 0:  # pragma: no cover
+            logger.warn("Unable to save last video! Did you call close()?")
