@@ -1,66 +1,77 @@
-import gym
+import gymnasium as gym
 import numpy as np
 import pytest
 import torch as th
+from gymnasium import spaces
 
 from stable_baselines3 import A2C, PPO, SAC
 from stable_baselines3.common.callbacks import BaseCallback
+from stable_baselines3.common.env_checker import check_env
 from stable_baselines3.common.policies import ActorCriticPolicy
 
 
 class CustomEnv(gym.Env):
     def __init__(self, max_steps=8):
-        super(CustomEnv, self).__init__()
-        self.observation_space = gym.spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32)
-        self.action_space = gym.spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32)
+        super().__init__()
+        self.observation_space = spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32)
+        self.action_space = spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32)
         self.max_steps = max_steps
         self.n_steps = 0
 
     def seed(self, seed):
         self.observation_space.seed(seed)
 
-    def reset(self):
+    def reset(self, *, seed: int | None = None, options: dict | None = None):
+        if seed is not None:
+            self.observation_space.seed(seed)
         self.n_steps = 0
-        return self.observation_space.sample()
+        return self.observation_space.sample(), {}
 
     def step(self, action):
         self.n_steps += 1
 
-        done = False
+        terminated = truncated = False
         reward = 0.0
         if self.n_steps >= self.max_steps:
             reward = 1.0
-            done = True
+            terminated = True
+            # To simplify GAE computation checks,
+            # we do not consider truncation here.
+            # Truncations are checked in InfiniteHorizonEnv
+            truncated = False
 
-        return self.observation_space.sample(), reward, done, {}
+        return self.observation_space.sample(), reward, terminated, truncated, {}
 
 
 class InfiniteHorizonEnv(gym.Env):
     def __init__(self, n_states=4):
         super().__init__()
         self.n_states = n_states
-        self.observation_space = gym.spaces.Discrete(n_states)
-        self.action_space = gym.spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32)
+        self.observation_space = spaces.Discrete(n_states)
+        self.action_space = spaces.Box(low=-1, high=1, shape=(2,), dtype=np.float32)
         self.current_state = 0
 
-    def reset(self):
+    def reset(self, *, seed: int | None = None, options: dict | None = None):
+        if seed is not None:
+            super().reset(seed=seed)
+
         self.current_state = 0
-        return self.current_state
+        return self.current_state, {}
 
     def step(self, action):
         self.current_state = (self.current_state + 1) % self.n_states
-        return self.current_state, 1.0, False, {}
+        return self.current_state, 1.0, False, False, {}
 
 
 class CheckGAECallback(BaseCallback):
     def __init__(self):
-        super(CheckGAECallback, self).__init__(verbose=0)
+        super().__init__(verbose=0)
 
     def _on_rollout_end(self):
         buffer = self.model.rollout_buffer
         rollout_size = buffer.size()
 
-        max_steps = self.training_env.envs[0].max_steps
+        max_steps = self.training_env.envs[0].get_wrapper_attr("max_steps")
         gamma = self.model.gamma
         gae_lambda = self.model.gae_lambda
         value = self.model.policy.constant_value
@@ -99,7 +110,7 @@ class CustomPolicy(ActorCriticPolicy):
     """Custom Policy with a constant value function"""
 
     def __init__(self, *args, **kwargs):
-        super(CustomPolicy, self).__init__(*args, **kwargs)
+        super().__init__(*args, **kwargs)
         self.constant_value = 0.0
 
     def forward(self, obs, deterministic=False):
@@ -107,6 +118,12 @@ class CustomPolicy(ActorCriticPolicy):
         # Overwrite values with ones
         values = th.ones_like(values) * self.constant_value
         return actions, values, log_prob
+
+
+@pytest.mark.parametrize("env_cls", [CustomEnv, InfiniteHorizonEnv])
+def test_env(env_cls):
+    # Check the env used for testing
+    check_env(env_cls(), skip_render_check=True)
 
 
 @pytest.mark.parametrize("model_class", [A2C, PPO])
