@@ -1,7 +1,7 @@
-import gym
+import gymnasium as gym
 import numpy as np
 import pytest
-from gym import spaces
+from gymnasium import spaces
 
 from stable_baselines3.common.vec_env import DummyVecEnv, VecCheckNan
 
@@ -9,10 +9,10 @@ from stable_baselines3.common.vec_env import DummyVecEnv, VecCheckNan
 class NanAndInfEnv(gym.Env):
     """Custom Environment that raised NaNs and Infs"""
 
-    metadata = {"render.modes": ["human"]}
+    metadata = {"render_modes": ["human"]}
 
     def __init__(self):
-        super(NanAndInfEnv, self).__init__()
+        super().__init__()
         self.action_space = spaces.Box(low=-np.inf, high=np.inf, shape=(1,), dtype=np.float64)
         self.observation_space = spaces.Box(low=-np.inf, high=np.inf, shape=(1,), dtype=np.float64)
 
@@ -24,13 +24,13 @@ class NanAndInfEnv(gym.Env):
             obs = float("inf")
         else:
             obs = 0
-        return [obs], 0.0, False, {}
+        return [obs], 0.0, False, False, {}
 
     @staticmethod
-    def reset():
-        return [0.0]
+    def reset(seed=None):
+        return [0.0], {}
 
-    def render(self, mode="human", close=False):
+    def render(self):
         pass
 
 
