@@ -1,9 +1,10 @@
 import types
+import warnings
 
-import gym
+import gymnasium as gym
 import numpy as np
 import pytest
-from gym import spaces
+from gymnasium import spaces
 
 from stable_baselines3.common.env_checker import check_env
 from stable_baselines3.common.envs import (
@@ -27,20 +28,20 @@ ENV_CLASSES = [
 ]
 
 
-@pytest.mark.parametrize("env_id", ["CartPole-v0", "Pendulum-v0"])
+@pytest.mark.parametrize("env_id", ["CartPole-v1", "Pendulum-v1"])
 def test_env(env_id):
     """
-    Check that environmnent integrated in Gym pass the test.
+    Check that environment integrated in Gym pass the test.
 
     :param env_id: (str)
     """
     env = gym.make(env_id)
-    with pytest.warns(None) as record:
+    with warnings.catch_warnings(record=True) as record:
         check_env(env)
 
-    # Pendulum-v0 will produce a warning because the action space is
+    # Pendulum-v1 will produce a warning because the action space is
     # in [-2, 2] and not [-1, 1]
-    if env_id == "Pendulum-v0":
+    if env_id == "Pendulum-v1":
         assert len(record) == 1
     else:
         # The other environments must pass without warning
@@ -50,7 +51,7 @@ def test_env(env_id):
 @pytest.mark.parametrize("env_class", ENV_CLASSES)
 def test_custom_envs(env_class):
     env = env_class()
-    with pytest.warns(None) as record:
+    with warnings.catch_warnings(record=True) as record:
         check_env(env)
     # No warnings for custom envs
     assert len(record) == 0
@@ -68,11 +69,22 @@ def test_custom_envs(env_class):
 def test_bit_flipping(kwargs):
     # Additional tests for BitFlippingEnv
     env = BitFlippingEnv(**kwargs)
-    with pytest.warns(None) as record:
+    with warnings.catch_warnings(record=True) as record:
         check_env(env)
 
     # No warnings for custom envs
     assert len(record) == 0
+
+    # Remove a key, must throw an error
+    obs_space = env.observation_space.spaces["observation"]
+    del env.observation_space.spaces["observation"]
+    with pytest.raises(AssertionError):
+        check_env(env)
+
+    # Rename a key, must throw an error
+    env.observation_space.spaces["obs"] = obs_space
+    with pytest.raises(AssertionError):
+        check_env(env)
 
 
 def test_high_dimension_action_space():
@@ -86,7 +98,7 @@ def test_high_dimension_action_space():
 
     # Patch to avoid error
     def patched_step(_action):
-        return env.observation_space.sample(), 0.0, False, {}
+        return env.observation_space.sample(), 0.0, False, False, {}
 
     env.step = patched_step
     check_env(env)
@@ -109,16 +121,28 @@ def test_high_dimension_action_space():
         spaces.Dict({"position": spaces.Dict({"abs": spaces.Discrete(5), "rel": spaces.Discrete(2)})}),
         # Small image inside a dict
         spaces.Dict({"img": spaces.Box(low=0, high=255, shape=(32, 32, 3), dtype=np.uint8)}),
+        # Non zero start index
+        spaces.Discrete(3, start=-1),
+        # 2D MultiDiscrete
+        spaces.MultiDiscrete(np.array([[4, 4], [2, 3]])),
+        # Non zero start index (MultiDiscrete)
+        spaces.MultiDiscrete([4, 4], start=[1, 0]),
+        # Non zero start index inside a Dict
+        spaces.Dict({"obs": spaces.Discrete(3, start=1)}),
     ],
 )
 def test_non_default_spaces(new_obs_space):
     env = FakeImageEnv()
     env.observation_space = new_obs_space
+
     # Patch methods to avoid errors
-    env.reset = new_obs_space.sample
+    def patched_reset(seed=None):
+        return new_obs_space.sample(), {}
+
+    env.reset = patched_reset
 
     def patched_step(_action):
-        return new_obs_space.sample(), 0.0, False, {}
+        return new_obs_space.sample(), 0.0, False, False, {}
 
     env.step = patched_step
     with pytest.warns(UserWarning):
@@ -136,27 +160,51 @@ def test_non_default_spaces(new_obs_space):
         spaces.Box(low=-1000, high=1000, shape=(3,), dtype=np.float32),
         # Too small range
         spaces.Box(low=-0.1, high=0.1, shape=(2,), dtype=np.float32),
-        # Inverted boundaries
-        spaces.Box(low=1, high=-1, shape=(2,), dtype=np.float32),
         # Same boundaries
         spaces.Box(low=1, high=1, shape=(2,), dtype=np.float32),
+        # Unbounded action space
+        spaces.Box(low=-np.inf, high=1, shape=(2,), dtype=np.float32),
         # Almost good, except for one dim
         spaces.Box(low=np.array([-1, -1, -1]), high=np.array([1, 1, 0.99]), dtype=np.float32),
+        # Non zero start index
+        spaces.Discrete(3, start=-1),
+        # Non zero start index (MultiDiscrete)
+        spaces.MultiDiscrete([4, 4], start=[1, 0]),
+        # 2D MultiDiscrete
+        spaces.MultiDiscrete(np.array([[4, 4], [2, 3]])),
     ],
 )
 def test_non_default_action_spaces(new_action_space):
     env = FakeImageEnv(discrete=False)
     # Default, should pass the test
-    with pytest.warns(None) as record:
+    with warnings.catch_warnings(record=True) as record:
         check_env(env)
 
     # No warnings for custom envs
     assert len(record) == 0
+
     # Change the action space
     env.action_space = new_action_space
 
-    with pytest.warns(UserWarning):
-        check_env(env)
+    # Discrete action space
+    if isinstance(new_action_space, (spaces.Discrete, spaces.MultiDiscrete)):
+        with pytest.warns(UserWarning):
+            check_env(env)
+        return
+
+    low, high = new_action_space.low[0], new_action_space.high[0]
+    # Unbounded action space throws an error,
+    # the rest only warning
+    if not np.all(np.isfinite(env.action_space.low)):
+        with pytest.raises(AssertionError), pytest.warns(UserWarning):
+            check_env(env)
+    # numpy >= 1.21 raises a ValueError
+    elif int(np.__version__.split(".")[1]) >= 21 and (low > high):
+        with pytest.raises(ValueError), pytest.warns(UserWarning):
+            check_env(env)
+    else:
+        with pytest.warns(UserWarning):
+            check_env(env)
 
 
 def check_reset_assert_error(env, new_reset_return):
@@ -166,8 +214,8 @@ def check_reset_assert_error(env, new_reset_return):
     :param new_reset_return: (Any)
     """
 
-    def wrong_reset():
-        return new_reset_return
+    def wrong_reset(seed=None):
+        return new_reset_return, {}
 
     # Patch the reset method with a wrong one
     env.reset = wrong_reset
@@ -185,14 +233,38 @@ def test_common_failures_reset():
     # The observation is not a numpy array
     check_reset_assert_error(env, 1)
 
+    # Return only obs (gym < 0.26)
+    def wrong_reset(self, seed=None):
+        return env.observation_space.sample()
+
+    env.reset = types.MethodType(wrong_reset, env)
+    with pytest.raises(AssertionError):
+        check_env(env)
+
+    # No seed parameter (gym < 0.26)
+    def wrong_reset(self):
+        return env.observation_space.sample(), {}
+
+    env.reset = types.MethodType(wrong_reset, env)
+    with pytest.raises(TypeError):
+        check_env(env)
+
     # Return not only the observation
     check_reset_assert_error(env, (env.observation_space.sample(), False))
 
     env = SimpleMultiObsEnv()
-    obs = env.reset()
 
-    def wrong_reset(self):
-        return {"img": obs["img"], "vec": obs["img"]}
+    # Observation keys and observation space keys must match
+    wrong_obs = env.observation_space.sample()
+    wrong_obs.pop("img")
+    check_reset_assert_error(env, wrong_obs)
+    wrong_obs = {**env.observation_space.sample(), "extra_key": None}
+    check_reset_assert_error(env, wrong_obs)
+
+    obs, _ = env.reset()
+
+    def wrong_reset(self, seed=None):
+        return {"img": obs["img"], "vec": obs["img"]}, {}
 
     env.reset = types.MethodType(wrong_reset, env)
     with pytest.raises(AssertionError) as excinfo:
@@ -225,25 +297,38 @@ def test_common_failures_step():
     env = IdentityEnvBox()
 
     # Wrong shape for the observation
-    check_step_assert_error(env, (np.ones((4,)), 1.0, False, {}))
+    check_step_assert_error(env, (np.ones((4,)), 1.0, False, False, {}))
     # Obs is not a numpy array
-    check_step_assert_error(env, (1, 1.0, False, {}))
+    check_step_assert_error(env, (1, 1.0, False, False, {}))
 
     # Return a wrong reward
-    check_step_assert_error(env, (env.observation_space.sample(), np.ones(1), False, {}))
+    check_step_assert_error(env, (env.observation_space.sample(), np.ones(1), False, False, {}))
 
     # Info dict is not returned
-    check_step_assert_error(env, (env.observation_space.sample(), 0.0, False))
+    check_step_assert_error(env, (env.observation_space.sample(), 0.0, False, False))
+
+    # Truncated is not returned (gym < 0.26)
+    check_step_assert_error(env, (env.observation_space.sample(), 0.0, False, {}))
 
     # Done is not a boolean
-    check_step_assert_error(env, (env.observation_space.sample(), 0.0, 3.0, {}))
-    check_step_assert_error(env, (env.observation_space.sample(), 0.0, 1, {}))
+    check_step_assert_error(env, (env.observation_space.sample(), 0.0, 3.0, False, {}))
+    check_step_assert_error(env, (env.observation_space.sample(), 0.0, 1, False, {}))
+    # Truncated is not a boolean
+    check_step_assert_error(env, (env.observation_space.sample(), 0.0, False, 1.0, {}))
 
     env = SimpleMultiObsEnv()
-    obs = env.reset()
+
+    # Observation keys and observation space keys must match
+    wrong_obs = env.observation_space.sample()
+    wrong_obs.pop("img")
+    check_step_assert_error(env, (wrong_obs, 0.0, False, False, {}))
+    wrong_obs = {**env.observation_space.sample(), "extra_key": None}
+    check_step_assert_error(env, (wrong_obs, 0.0, False, False, {}))
+
+    obs, _ = env.reset()
 
     def wrong_step(self, action):
-        return {"img": obs["vec"], "vec": obs["vec"]}, 0.0, False, {}
+        return {"img": obs["vec"], "vec": obs["vec"]}, 0.0, False, False, {}
 
     env.step = types.MethodType(wrong_step, env)
     with pytest.raises(AssertionError) as excinfo:
