@@ -3,19 +3,18 @@ import pathlib
 import warnings
 from copy import deepcopy
 
-import gym
 import numpy as np
 import pytest
 import torch as th
 
 from stable_baselines3 import DDPG, DQN, SAC, TD3, HerReplayBuffer
+from stable_baselines3.common.env_util import make_vec_env
 from stable_baselines3.common.envs import BitFlippingEnv
 from stable_baselines3.common.evaluation import evaluate_policy
 from stable_baselines3.common.monitor import Monitor
 from stable_baselines3.common.noise import NormalActionNoise
-from stable_baselines3.common.vec_env import DummyVecEnv
+from stable_baselines3.common.vec_env import SubprocVecEnv
 from stable_baselines3.her.goal_selection_strategy import GoalSelectionStrategy
-from stable_baselines3.her.her_replay_buffer import get_time_limit
 
 
 def test_import_error():
@@ -27,18 +26,22 @@ def test_import_error():
 
 
 @pytest.mark.parametrize("model_class", [SAC, TD3, DDPG, DQN])
-@pytest.mark.parametrize("online_sampling", [True, False])
 @pytest.mark.parametrize("image_obs_space", [True, False])
-def test_her(model_class, online_sampling, image_obs_space):
+def test_her(model_class, image_obs_space):
     """
     Test Hindsight Experience Replay.
     """
+    n_envs = 1
     n_bits = 4
-    env = BitFlippingEnv(
-        n_bits=n_bits,
-        continuous=not (model_class == DQN),
-        image_obs_space=image_obs_space,
-    )
+
+    def env_fn():
+        return BitFlippingEnv(
+            n_bits=n_bits,
+            continuous=not (model_class == DQN),
+            image_obs_space=image_obs_space,
+        )
+
+    env = make_vec_env(env_fn, n_envs)
 
     model = model_class(
         "MultiInputPolicy",
@@ -47,18 +50,28 @@ def test_her(model_class, online_sampling, image_obs_space):
         replay_buffer_kwargs=dict(
             n_sampled_goal=2,
             goal_selection_strategy="future",
-            online_sampling=online_sampling,
-            max_episode_length=n_bits,
+            copy_info_dict=True,
         ),
         train_freq=4,
-        gradient_steps=1,
+        gradient_steps=n_envs,
         policy_kwargs=dict(net_arch=[64]),
         learning_starts=100,
         buffer_size=int(2e4),
     )
 
     model.learn(total_timesteps=150)
-    evaluate_policy(model, Monitor(env))
+    evaluate_policy(model, Monitor(env_fn()))
+
+
+@pytest.mark.parametrize("model_class", [TD3, DQN])
+@pytest.mark.parametrize("image_obs_space", [True, False])
+def test_multiprocessing(model_class, image_obs_space):
+    def env_fn():
+        return BitFlippingEnv(n_bits=4, continuous=not (model_class == DQN), image_obs_space=image_obs_space)
+
+    env = make_vec_env(env_fn, n_envs=2, vec_env_cls=SubprocVecEnv)
+    model = model_class("MultiInputPolicy", env, replay_buffer_class=HerReplayBuffer, buffer_size=int(2e4), train_freq=4)
+    model.learn(total_timesteps=150)
 
 
 @pytest.mark.parametrize(
@@ -72,12 +85,16 @@ def test_her(model_class, online_sampling, image_obs_space):
         GoalSelectionStrategy.FUTURE,
     ],
 )
-@pytest.mark.parametrize("online_sampling", [True, False])
-def test_goal_selection_strategy(goal_selection_strategy, online_sampling):
+def test_goal_selection_strategy(goal_selection_strategy):
     """
     Test different goal strategies.
     """
-    env = BitFlippingEnv(continuous=True)
+    n_envs = 2
+
+    def env_fn():
+        return BitFlippingEnv(continuous=True)
+
+    env = make_vec_env(env_fn, n_envs)
 
     normal_action_noise = NormalActionNoise(np.zeros(1), 0.1 * np.ones(1))
 
@@ -87,12 +104,10 @@ def test_goal_selection_strategy(goal_selection_strategy, online_sampling):
         replay_buffer_class=HerReplayBuffer,
         replay_buffer_kwargs=dict(
             goal_selection_strategy=goal_selection_strategy,
-            online_sampling=online_sampling,
-            max_episode_length=10,
             n_sampled_goal=2,
         ),
         train_freq=4,
-        gradient_steps=1,
+        gradient_steps=n_envs,
         policy_kwargs=dict(net_arch=[64]),
         learning_starts=100,
         buffer_size=int(1e5),
@@ -104,16 +119,20 @@ def test_goal_selection_strategy(goal_selection_strategy, online_sampling):
 
 @pytest.mark.parametrize("model_class", [SAC, TD3, DDPG, DQN])
 @pytest.mark.parametrize("use_sde", [False, True])
-@pytest.mark.parametrize("online_sampling", [False, True])
-def test_save_load(tmp_path, model_class, use_sde, online_sampling):
+def test_save_load(tmp_path, model_class, use_sde):
     """
     Test if 'save' and 'load' saves and loads model correctly
     """
     if use_sde and model_class != SAC:
         pytest.skip("Only SAC has gSDE support")
 
+    n_envs = 2
     n_bits = 4
-    env = BitFlippingEnv(n_bits=n_bits, continuous=not (model_class == DQN))
+
+    def env_fn():
+        return BitFlippingEnv(n_bits=n_bits, continuous=not (model_class == DQN))
+
+    env = make_vec_env(env_fn, n_envs)
 
     kwargs = dict(use_sde=True) if use_sde else {}
 
@@ -125,8 +144,6 @@ def test_save_load(tmp_path, model_class, use_sde, online_sampling):
         replay_buffer_kwargs=dict(
             n_sampled_goal=2,
             goal_selection_strategy="future",
-            online_sampling=online_sampling,
-            max_episode_length=n_bits,
         ),
         verbose=0,
         tau=0.05,
@@ -135,7 +152,7 @@ def test_save_load(tmp_path, model_class, use_sde, online_sampling):
         policy_kwargs=dict(net_arch=[64]),
         buffer_size=int(1e5),
         gamma=0.98,
-        gradient_steps=1,
+        gradient_steps=n_envs,
         train_freq=4,
         learning_starts=100,
         **kwargs
@@ -143,20 +160,15 @@ def test_save_load(tmp_path, model_class, use_sde, online_sampling):
 
     model.learn(total_timesteps=150)
 
-    obs = env.reset()
-
-    observations = {key: [] for key in obs.keys()}
-    for _ in range(10):
-        obs = env.step(env.action_space.sample())[0]
-        for key in obs.keys():
-            observations[key].append(obs[key])
-    observations = {key: np.array(obs) for key, obs in observations.items()}
+    env.reset()
+    action = np.array([env.action_space.sample() for _ in range(n_envs)])
+    observations = env.step(action)[0]
 
     # Get dictionary of current parameters
     params = deepcopy(model.policy.state_dict())
 
     # Modify all parameters to be random values
-    random_params = dict((param_name, th.rand_like(param)) for param_name, param in params.items())
+    random_params = {param_name: th.rand_like(param) for param_name, param in params.items()}
 
     # Update model parameters with the new random values
     model.policy.load_state_dict(random_params)
@@ -210,9 +222,9 @@ def test_save_load(tmp_path, model_class, use_sde, online_sampling):
     os.remove(tmp_path / "test_save.zip")
 
 
-@pytest.mark.parametrize("online_sampling", [False, True])
+@pytest.mark.parametrize("n_envs", [1, 2])
 @pytest.mark.parametrize("truncate_last_trajectory", [False, True])
-def test_save_load_replay_buffer(tmp_path, recwarn, online_sampling, truncate_last_trajectory):
+def test_save_load_replay_buffer(n_envs, tmp_path, recwarn, truncate_last_trajectory):
     """
     Test if 'save_replay_buffer' and 'load_replay_buffer' works correctly
     """
@@ -222,7 +234,11 @@ def test_save_load_replay_buffer(tmp_path, recwarn, online_sampling, truncate_la
 
     path = pathlib.Path(tmp_path / "replay_buffer.pkl")
     path.parent.mkdir(exist_ok=True, parents=True)  # to not raise a warning
-    env = BitFlippingEnv(n_bits=4, continuous=True)
+
+    def env_fn():
+        return BitFlippingEnv(n_bits=4, continuous=True)
+
+    env = make_vec_env(env_fn, n_envs)
     model = SAC(
         "MultiInputPolicy",
         env,
@@ -230,68 +246,43 @@ def test_save_load_replay_buffer(tmp_path, recwarn, online_sampling, truncate_la
         replay_buffer_kwargs=dict(
             n_sampled_goal=2,
             goal_selection_strategy="future",
-            online_sampling=online_sampling,
-            max_episode_length=4,
         ),
-        gradient_steps=1,
+        gradient_steps=n_envs,
         train_freq=4,
         buffer_size=int(2e4),
         policy_kwargs=dict(net_arch=[64]),
-        seed=1,
+        seed=0,
     )
     model.learn(200)
-    if online_sampling:
-        old_replay_buffer = deepcopy(model.replay_buffer)
-    else:
-        old_replay_buffer = deepcopy(model.replay_buffer.replay_buffer)
+    old_replay_buffer = deepcopy(model.replay_buffer)
+
     model.save_replay_buffer(path)
     del model.replay_buffer
 
     with pytest.raises(AttributeError):
-        model.replay_buffer
+        model.replay_buffer  # noqa: B018
 
     # Check that there is no warning
     assert len(recwarn) == 0
 
     model.load_replay_buffer(path, truncate_last_traj=truncate_last_trajectory)
 
-    if truncate_last_trajectory:
+    if truncate_last_trajectory and (old_replay_buffer.dones[old_replay_buffer.pos - 1] == 0).any():
         assert len(recwarn) == 1
         warning = recwarn.pop(UserWarning)
         assert "The last trajectory in the replay buffer will be truncated" in str(warning.message)
     else:
         assert len(recwarn) == 0
 
-    if online_sampling:
-        n_episodes_stored = model.replay_buffer.n_episodes_stored
-        assert np.allclose(
-            old_replay_buffer._buffer["observation"][:n_episodes_stored],
-            model.replay_buffer._buffer["observation"][:n_episodes_stored],
-        )
-        assert np.allclose(
-            old_replay_buffer._buffer["next_obs"][:n_episodes_stored],
-            model.replay_buffer._buffer["next_obs"][:n_episodes_stored],
-        )
-        assert np.allclose(
-            old_replay_buffer._buffer["action"][:n_episodes_stored],
-            model.replay_buffer._buffer["action"][:n_episodes_stored],
-        )
-        assert np.allclose(
-            old_replay_buffer._buffer["reward"][:n_episodes_stored],
-            model.replay_buffer._buffer["reward"][:n_episodes_stored],
-        )
-        # we might change the last done of the last trajectory so we don't compare it
-        assert np.allclose(
-            old_replay_buffer._buffer["done"][: n_episodes_stored - 1],
-            model.replay_buffer._buffer["done"][: n_episodes_stored - 1],
-        )
-    else:
-        replay_buffer = model.replay_buffer.replay_buffer
-        assert np.allclose(old_replay_buffer.observations["observation"], replay_buffer.observations["observation"])
-        assert np.allclose(old_replay_buffer.observations["desired_goal"], replay_buffer.observations["desired_goal"])
-        assert np.allclose(old_replay_buffer.actions, replay_buffer.actions)
-        assert np.allclose(old_replay_buffer.rewards, replay_buffer.rewards)
-        assert np.allclose(old_replay_buffer.dones, replay_buffer.dones)
+    replay_buffer = model.replay_buffer
+    pos = replay_buffer.pos
+    for key in ["observation", "desired_goal", "achieved_goal"]:
+        assert np.allclose(old_replay_buffer.observations[key][:pos], replay_buffer.observations[key][:pos])
+        assert np.allclose(old_replay_buffer.next_observations[key][:pos], replay_buffer.next_observations[key][:pos])
+    assert np.allclose(old_replay_buffer.actions[:pos], replay_buffer.actions[:pos])
+    assert np.allclose(old_replay_buffer.rewards[:pos], replay_buffer.rewards[:pos])
+    # we might change the last done of the last trajectory so we don't compare it
+    assert np.allclose(old_replay_buffer.dones[: pos - 1], replay_buffer.dones[: pos - 1])
 
     # test if continuing training works properly
     reset_num_timesteps = False if truncate_last_trajectory is False else True
@@ -304,7 +295,12 @@ def test_full_replay_buffer():
     It should not sample the current episode which is not finished.
     """
     n_bits = 4
-    env = BitFlippingEnv(n_bits=n_bits, continuous=True)
+    n_envs = 2
+
+    def env_fn():
+        return BitFlippingEnv(n_bits=n_bits, continuous=True)
+
+    env = make_vec_env(env_fn, n_envs)
 
     # use small buffer size to get the buffer full
     model = SAC(
@@ -314,14 +310,12 @@ def test_full_replay_buffer():
         replay_buffer_kwargs=dict(
             n_sampled_goal=2,
             goal_selection_strategy="future",
-            online_sampling=True,
-            max_episode_length=n_bits,
         ),
         gradient_steps=1,
         train_freq=4,
         policy_kwargs=dict(net_arch=[64]),
-        learning_starts=1,
-        buffer_size=20,
+        learning_starts=n_bits * n_envs,
+        buffer_size=20 * n_envs,
         verbose=1,
         seed=757,
     )
@@ -329,49 +323,126 @@ def test_full_replay_buffer():
     model.learn(total_timesteps=100)
 
 
-def test_get_max_episode_length():
-    dict_env = DummyVecEnv([lambda: BitFlippingEnv()])
+@pytest.mark.parametrize("n_envs", [1, 2])
+@pytest.mark.parametrize("n_steps", [4, 5])
+@pytest.mark.parametrize("handle_timeout_termination", [False, True])
+def test_truncate_last_trajectory(n_envs, recwarn, n_steps, handle_timeout_termination):
+    """
+    Test if 'truncate_last_trajectory' works correctly
+    """
+    # remove gym warnings
+    warnings.filterwarnings(action="ignore", category=DeprecationWarning)
+    warnings.filterwarnings(action="ignore", category=UserWarning, module="gym")
 
-    # Cannot infer max epsiode length
-    with pytest.raises(ValueError):
-        get_time_limit(dict_env, current_max_episode_length=None)
+    n_bits = 4
 
-    default_length = 10
-    assert get_time_limit(dict_env, current_max_episode_length=default_length) == default_length
+    def env_fn():
+        return BitFlippingEnv(n_bits=n_bits, continuous=True)
 
-    env = gym.make("CartPole-v1")
-    vec_env = DummyVecEnv([lambda: env])
+    venv = make_vec_env(env_fn, n_envs)
 
-    assert get_time_limit(vec_env, current_max_episode_length=None) == 500
-    # Overwrite max_episode_steps
-    assert get_time_limit(vec_env, current_max_episode_length=default_length) == default_length
+    replay_buffer = HerReplayBuffer(
+        buffer_size=int(1e4),
+        observation_space=venv.observation_space,
+        action_space=venv.action_space,
+        env=venv,
+        n_envs=n_envs,
+        n_sampled_goal=2,
+        goal_selection_strategy="future",
+    )
 
-    # Set max_episode_steps to None
-    env.spec.max_episode_steps = None
-    vec_env = DummyVecEnv([lambda: env])
-    with pytest.raises(ValueError):
-        get_time_limit(vec_env, current_max_episode_length=None)
+    observations = venv.reset()
+    for _ in range(n_steps):
+        actions = np.random.rand(n_envs, n_bits)
+        next_observations, rewards, dones, infos = venv.step(actions)
+        replay_buffer.add(observations, next_observations, actions, rewards, dones, infos)
+        observations = next_observations
 
-    # Initialize HER and specify max_episode_length, should not raise an issue
-    DQN("MultiInputPolicy", dict_env, replay_buffer_class=HerReplayBuffer, replay_buffer_kwargs=dict(max_episode_length=5))
+    old_replay_buffer = deepcopy(replay_buffer)
+    pos = replay_buffer.pos
+    if handle_timeout_termination:
+        env_idx_not_finished = np.where(replay_buffer._current_ep_start != pos)[0]
 
-    with pytest.raises(ValueError):
-        DQN("MultiInputPolicy", dict_env, replay_buffer_class=HerReplayBuffer)
+    # Check that there is no warning
+    assert len(recwarn) == 0
 
-    # Wrapped in a timelimit, should be fine
-    # Note: it requires env.spec to be defined
-    env = DummyVecEnv([lambda: gym.wrappers.TimeLimit(BitFlippingEnv(), 10)])
-    DQN("MultiInputPolicy", env, replay_buffer_class=HerReplayBuffer, replay_buffer_kwargs=dict(max_episode_length=5))
+    replay_buffer.truncate_last_trajectory()
+
+    if (old_replay_buffer.dones[pos - 1] == 0).any():
+        # at least one episode in the replay buffer did not finish
+        assert len(recwarn) == 1
+        warning = recwarn.pop(UserWarning)
+        assert "The last trajectory in the replay buffer will be truncated" in str(warning.message)
+    else:
+        # all episodes in the replay buffer are finished
+        assert len(recwarn) == 0
+
+    # next episode starts at current pos
+    assert (replay_buffer._current_ep_start == pos).all()
+    # done = True for last episodes
+    assert (replay_buffer.dones[pos - 1] == 1).all()
+    # for all episodes that are not finished before truncate_last_trajectory: timeouts should be 1
+    if handle_timeout_termination:
+        assert (replay_buffer.timeouts[pos - 1, env_idx_not_finished] == 1).all()
+    # episode length should be != 0 -> episode can be sampled
+    assert (replay_buffer.ep_length[pos - 1] != 0).all()
+
+    # replay buffer should not have changed after truncate_last_trajectory (except dones[pos-1])
+    for key in ["observation", "desired_goal", "achieved_goal"]:
+        assert np.allclose(old_replay_buffer.observations[key], replay_buffer.observations[key])
+        assert np.allclose(old_replay_buffer.next_observations[key], replay_buffer.next_observations[key])
+    assert np.allclose(old_replay_buffer.actions, replay_buffer.actions)
+    assert np.allclose(old_replay_buffer.rewards, replay_buffer.rewards)
+    # we might change the last done of the last trajectory so we don't compare it
+    assert np.allclose(old_replay_buffer.dones[: pos - 1], replay_buffer.dones[: pos - 1])
+    assert np.allclose(old_replay_buffer.dones[pos:], replay_buffer.dones[pos:])
+
+    for _ in range(10):
+        actions = np.random.rand(n_envs, n_bits)
+        next_observations, rewards, dones, infos = venv.step(actions)
+        replay_buffer.add(observations, next_observations, actions, rewards, dones, infos)
+        observations = next_observations
+
+    # old oberservations must remain unchanged
+    for key in ["observation", "desired_goal", "achieved_goal"]:
+        assert np.allclose(old_replay_buffer.observations[key][:pos], replay_buffer.observations[key][:pos])
+        assert np.allclose(old_replay_buffer.next_observations[key][:pos], replay_buffer.next_observations[key][:pos])
+    assert np.allclose(old_replay_buffer.actions[:pos], replay_buffer.actions[:pos])
+    assert np.allclose(old_replay_buffer.rewards[:pos], replay_buffer.rewards[:pos])
+    assert np.allclose(old_replay_buffer.dones[: pos - 1], replay_buffer.dones[: pos - 1])
+
+    # new oberservations must differ from old observations
+    end_pos = replay_buffer.pos
+    for key in ["observation", "desired_goal", "achieved_goal"]:
+        assert not np.allclose(old_replay_buffer.observations[key][pos:end_pos], replay_buffer.observations[key][pos:end_pos])
+        assert not np.allclose(
+            old_replay_buffer.next_observations[key][pos:end_pos], replay_buffer.next_observations[key][pos:end_pos]
+        )
+    assert not np.allclose(old_replay_buffer.actions[pos:end_pos], replay_buffer.actions[pos:end_pos])
+    assert not np.allclose(old_replay_buffer.rewards[pos:end_pos], replay_buffer.rewards[pos:end_pos])
+    assert not np.allclose(old_replay_buffer.dones[pos - 1 : end_pos], replay_buffer.dones[pos - 1 : end_pos])
+
+    # all entries with index >= replay_buffer.pos must remain unchanged
+    for key in ["observation", "desired_goal", "achieved_goal"]:
+        assert np.allclose(old_replay_buffer.observations[key][end_pos:], replay_buffer.observations[key][end_pos:])
+        assert np.allclose(old_replay_buffer.next_observations[key][end_pos:], replay_buffer.next_observations[key][end_pos:])
+    assert np.allclose(old_replay_buffer.actions[end_pos:], replay_buffer.actions[end_pos:])
+    assert np.allclose(old_replay_buffer.rewards[end_pos:], replay_buffer.rewards[end_pos:])
+    assert np.allclose(old_replay_buffer.dones[end_pos:], replay_buffer.dones[end_pos:])
 
 
-@pytest.mark.parametrize("online_sampling", [False, True])
 @pytest.mark.parametrize("n_bits", [10])
-def test_performance_her(online_sampling, n_bits):
+def test_performance_her(n_bits):
     """
     That DQN+HER can solve BitFlippingEnv.
     It should not work when n_sampled_goal=0 (DQN alone).
     """
-    env = BitFlippingEnv(n_bits=n_bits, continuous=False)
+    n_envs = 2
+
+    def env_fn():
+        return BitFlippingEnv(n_bits=n_bits, continuous=False)
+
+    env = make_vec_env(env_fn, n_envs)
 
     model = DQN(
         "MultiInputPolicy",
@@ -380,12 +451,11 @@ def test_performance_her(online_sampling, n_bits):
         replay_buffer_kwargs=dict(
             n_sampled_goal=5,
             goal_selection_strategy="future",
-            online_sampling=online_sampling,
-            max_episode_length=n_bits,
         ),
         verbose=1,
         learning_rate=5e-4,
         train_freq=1,
+        gradient_steps=n_envs,
         learning_starts=100,
         exploration_final_eps=0.02,
         target_update_interval=500,
